+++ RockPaperScissors/RockPaperScissors/Enum/MukChiPPaOption.swift
@@ -1,0 +1,15 @@
+//
+//  MukChiPPaOption.swift
+//  RockPaperScissors
+//
+//  Created by 권태호 on 29/11/2023.
+//
+
+import Foundation
+
+enum MukChiPPaOption: String {
+    case muk = "1"
+    case chi = "2"
+    case ppa = "3"
+    case exit = "게임종료"
+}
