+++ RockPaperScissors/RockPaperScissors/Enum/UserMenuOption.swift
@@ -1,0 +1,15 @@
+//
+//  UserMenuOption.swift
+//  RockPaperScissors
+//
+//  Created by 권태호 on 29/11/2023.
+//
+
+import Foundation
+
+enum UserMenuOption: String {
+    case scissor = "1"
+    case rock = "2"
+    case paper = "3"
+    case exit = "0"
+}
