+++ RockPaperScissors/RockPaperScissors/mukcchippaGameManager.swift
@@ -1,10 +1,8 @@
 //
-//  safa.swift
+//  mukcchippaGameManager.swift
 //  RockPaperScissors
 //
 //  Created by 권태호 on 29/11/2023.
 //
 
 import Foundation
-
-
