+++ RockPaperScissors/RockPaperScissors/rockSicssorPaperManager.swift
@@ -1,0 +1,13 @@
+//
+//  rockSicssorPaperManager.swift
+//  RockPaperScissors
+//
+//  Created by 권태호 on 29/11/2023.
+//
+
+import Foundation
+
+struct rockSicssorPaperGameManager {
+    
+    //가위바위보 로직 구현
+}
